refactor(taco): remove redundant exports from taco package
