+++ examples/taco/webpack-5/src/index.ts
@@ -37,7 +37,7 @@ const runExample = async () => {
 
   console.log('Encrypting message...');
   const message = toBytes('this is a secret');
-  const hasPositiveBalance = new conditions.RpcCondition({
+  const hasPositiveBalance = new conditions.base.RpcCondition({
     chain: 80001,
     method: 'eth_getBalance',
     parameters: [':userAddress', 'latest'],
