+++ packages/taco/examples/conditions.ts
@@ -78,7 +78,7 @@ console.assert(
   'CompoundCondition requires signer',
 );
 
-const myFunctionAbi: conditions.FunctionAbiProps = {
+const myFunctionAbi: conditions.base.FunctionAbiProps = {
   name: 'myFunction',
   type: 'function',
   stateMutability: 'view',
