+++ packages/taco/test/taco.test.ts
@@ -28,7 +28,7 @@ import {
 
 // Shared test variables
 const message = 'this is a secret';
-const ownsNFT = new conditions.ERC721Ownership({
+const ownsNFT = new conditions.predefined.ERC721Ownership({
   contractAddress: '0x1e988ba4692e52Bc50b375bcC8585b95c48AaD77',
   parameters: [3591],
   chain: 5,
