+++ packages/taco/test/test-utils.ts
@@ -36,17 +36,16 @@ import { SpyInstance, vi } from 'vitest';
 
 import {
   ConditionExpression,
-  ContractConditionProps,
-  ContractConditionType,
-  ERC721Balance,
-  FunctionAbiProps,
+} from '../src/conditions';
+import {
+  ContractConditionProps, ContractConditionType, FunctionAbiProps,
   ReturnValueTestProps,
-  RpcConditionProps,
-  RpcConditionType,
+  RpcConditionProps, RpcConditionType,
   TimeConditionMethod,
   TimeConditionProps,
-  TimeConditionType,
-} from '../src/conditions';
+  TimeConditionType
+} from "../src/conditions/base";
+import {ERC721Balance} from "../src/conditions/predefined";
 import { DkgClient, DkgRitual } from '../src/dkg';
 import { encryptMessage } from '../src/tdec';
